refactor(webui): isolate desktop application boundary
