+++ frontend/chat/src/mobile-pairing.test.mjs
@@ -4,7 +4,7 @@ import test from "node:test";
 
 import { parsePairedDevice, parsePairingOffer, parsePairingStatus } from "./mobile-pairing-data.ts";
 
-const entry = await readFile(new URL("./main.tsx", import.meta.url), "utf8");
+const app = await readFile(new URL("./desktop-chat-app.tsx", import.meta.url), "utf8");
 const dialog = await readFile(new URL("./mobile-pairing-dialog.tsx", import.meta.url), "utf8");
 const controller = await readFile(new URL("./use-mobile-pairing.ts", import.meta.url), "utf8");
 
@@ -40,8 +40,8 @@ test("pairing view, controller, and protocol have separate owners", () => {
   assert.doesNotMatch(dialog, /\bfetch\b|toDataURL|parsePairingOffer/);
   assert.match(controller, /new AbortController\(\)/);
   assert.match(controller, /actionRef\.current\?\.abort\(\)/);
-  assert.match(entry, /const LazyMobilePairingDialog = lazy/);
-  assert.doesNotMatch(entry, /import \{ MobilePairingDialog \} from/);
+  assert.match(app, /const LazyMobilePairingDialog = lazy/);
+  assert.doesNotMatch(app, /import \{ MobilePairingDialog \} from/);
 });
 
 test("pairing countdown is visible without announcing every second", () => {
